+++ src/ops.jl
@@ -75,6 +75,7 @@ unary_ops = [
 # "j1",
 ("gamma_impl", "gamma"),
 ("lgamma", "loggamma"), # lgamma(x::Real)` is deprecated, use `(logabsgamma(x))[1]` instead. Other alternative is loggamma, throws a DomainError if gamma(x) is negative.
+("lgamma", "lgamma"), # TODO: deprecate after everyone switches to SpecialFunctions 0.8
 ("digamma_impl", "digamma"),
 ("trigamma_impl", "trigamma"),
 #